updating breadcrumbs to show title of item page
