+++ app/routes/_layout+/items.$itemId.tsx
@@ -1,6 +1,6 @@
 import type { ActionArgs, LoaderArgs, V2_MetaFunction } from "@remix-run/node";
 import { redirect, json } from "@remix-run/node";
-import { Form, useCatch, useLoaderData } from "@remix-run/react";
+import { Form, useCatch, useLoaderData, useMatches } from "@remix-run/react";
 import type { HeaderData } from "~/components/layout/header/types";
 
 import { requireAuthSession, commitAuthSession } from "~/modules/auth";
@@ -30,6 +30,10 @@ export const meta: V2_MetaFunction<typeof loader> = ({ data }) => [
   { title: appendToMetaTitle(data.header.title) },
 ];
 
+export const handle = {
+  breadcrumb: () => "single",
+};
+
 export async function action({ request, params }: ActionArgs) {
   assertIsDelete(request);
   const id = getRequiredParam(params, "itemId");
@@ -46,6 +50,9 @@ export async function action({ request, params }: ActionArgs) {
 
 export default function ItemDetailsPage() {
   const data = useLoaderData<typeof loader>();
+  const m = useMatches();
+
+  console.log(m);
 
   return (
     <div>
